+++ pyro/contrib/funsor/handlers/__init__.py
@@ -9,8 +9,8 @@ from pyro.poutine import (  # noqa: F401
 )
 
 
-from .enum_messenger import EnumMessenger, MarkovMessenger, queue  # noqa: F401
-from .named_messenger import NamedMessenger
+from .enum_messenger import EnumMessenger, queue  # noqa: F401
+from .named_messenger import MarkovMessenger, NamedMessenger
 from .plate_messenger import PlateMessenger
 from .replay_messenger import ReplayMessenger
 from .trace_messenger import TraceMessenger
